moved files to UI and expenses folder
