+++ src/components/Expenses/Expenses.js
@@ -1,6 +1,6 @@
 import ExpenseItem from "./ExpenseItems";
 import "./Expenses.css";
-import Card from "./Card";
+import Card from "../UI/Card";
 
 function Expenses(props) {
   return (
